refactor: Rename file to match content
