+++ liquid-interpreter/src/lib.rs
@@ -23,7 +23,7 @@ mod argument;
 mod context;
 mod filter;
 mod globals;
-mod output;
+mod filter_chain;
 mod renderable;
 mod template;
 mod text;
@@ -33,7 +33,7 @@ pub use self::argument::*;
 pub use self::context::*;
 pub use self::filter::*;
 pub use self::globals::*;
-pub use self::output::*;
+pub use self::filter_chain::*;
 pub use self::renderable::*;
 pub use self::template::*;
 pub use self::text::*;
